+++ Sources/Sentry/SentryHttpTransport.m
@@ -249,6 +249,7 @@
     [self recordLostEvent:dataCategory reason:SentryDiscardReasonRateLimitBackoff];
     [self recordLostSpans:envelopeItem reason:SentryDiscardReasonRateLimitBackoff];
     [self recordLostLogBytes:envelopeItem reason:SentryDiscardReasonRateLimitBackoff];
+    [self recordLostTraceMetricBytes:envelopeItem reason:SentryDiscardReasonRateLimitBackoff];
 }
 
 - (void)envelopeItemDeleted:(SentryEnvelopeItem *)envelopeItem
@@ -257,6 +258,7 @@
     [self recordLostEvent:dataCategory reason:SentryDiscardReasonCacheOverflow];
     [self recordLostSpans:envelopeItem reason:SentryDiscardReasonCacheOverflow];
     [self recordLostLogBytes:envelopeItem reason:SentryDiscardReasonCacheOverflow];
+    [self recordLostTraceMetricBytes:envelopeItem reason:SentryDiscardReasonCacheOverflow];
 }
 
 #pragma mark private methods
@@ -471,6 +473,7 @@
         [self recordLostEvent:category reason:SentryDiscardReasonSendError];
         [self recordLostSpans:item reason:SentryDiscardReasonSendError];
         [self recordLostLogBytes:item reason:SentryDiscardReasonSendError];
+        [self recordLostTraceMetricBytes:item reason:SentryDiscardReasonSendError];
     }
 }
 
@@ -479,6 +482,15 @@
     if ([SentryEnvelopeItemTypes.log isEqualToString:envelopeItem.type]) {
         NSUInteger byteCount = envelopeItem.data.length;
         [self recordLostEvent:SentryDataCategoryLogByte reason:reason quantity:byteCount];
+    }
+}
+
+- (void)recordLostTraceMetricBytes:(SentryEnvelopeItem *)envelopeItem
+                            reason:(SentryDiscardReason)reason
+{
+    if ([SentryEnvelopeItemTypes.traceMetric isEqualToString:envelopeItem.type]) {
+        NSUInteger byteCount = envelopeItem.data.length;
+        [self recordLostEvent:SentryDataCategoryTraceMetricByte reason:reason quantity:byteCount];
     }
 }